requirements.txt
pyxel>=1.9.0
python-vlc>=3.0.0
